backend/enhanced_requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
streamlit
opencv-python
numpy
pillow
pytesseract
reportlab
pandas
python-dateutil
easyocr
